core: add Workspace back in
The Workspace turns out to be useful, particularly when handled
correctly. This allows for a unification of all packages.

Additionally, since we now are focusing on applications and servers now,
the decision of which packages to build by default becomes clearer.

Finally, we'll be able to use `xbuild` to construct more complex
workflows with a full Workspace.

Signed-off-by: Sean Cross <sean@xobs.io>