loader: fix double-allocation of memory regions
Due to an error in how small regions were loaded, memory was mapped to
the wrong offset. This didn't affect us for the most part, since these
small regions were mostly in the backtrace report which we don't usually
hit.

With this patch, we now can verify that processes are loaded correctly.
For example:

    (gdb) compare-sections
    Section .rodata, range 0x100e0 -- 0x13bdc: matched.
    Section .eh_frame, range 0x13bdc -- 0x15040: matched.
    Section .text, range 0x16040 -- 0x2dd6e: matched.
    Section .sdata, range 0x2ed70 -- 0x2ed78: matched.
    Section .data, range 0x2ed78 -- 0x2eda4: matched.
    (gdb)

Signed-off-by: Sean Cross <sean@xobs.io>