a patch for issue #73
the problem is that during the computation of the lines, the
cursor isn't wrapping to the next line if the last character
would render "into" the edge of the box. adding a trailing
space to any incoming strings forces the line to wrap.

you can aggravate this condition by typing the following
sentence into the shell:

this is a test of a message that is too long a message

when it fails, it renders

this is a test of a messag
e that is too long a messag

adding the trailing space causes it to render as

this is a test of a messag
e that is too long a messa
ge

which is the desired behavior