that was a long walk to fix a bad architectural decision
trying to allocate key indicies within the dictionary pool in a late-binding
way is hard to do in rust, because of various interior mutability problems.
i guess it could be done but it turned out to be easier to pro-actively
track everything and generate them on demand.

allocations are hard in rust. i guess that's why the alloc thing is
still not stable.